Bug 457621 - changelog Atom feed is hard to read with long lines wrapped in <pre>, r=ted
We should make both multi-line and single-line commit messages work in changelog Atom feeds, no matter how long the single line is.

The current approach, using inline (type="xhtml") content which wraps the plain-text commit message (with any markup-significant characters it might have contained escaped) in a <pre> XHTML element, fails badly in feed readers which obey it whenever a single line message exceeds the width of the feed reader's window - horizontal scrollbars are the most painful form of reading.

The naive approach, "we've got plain text data, Atom has a type="text", we should use that" fails for multiline messages, since it depends on all consumers using an XML parser which properly supports xml:space="preserve" and then realizing that they need to preserve the whitespace between their parser and their display, a pair of not-so-good dependencies.

This patch's approach is to escape any less-than, greater-than, and ampersand characters in the original (the first "escape" filter), then convert newlines to <br> (the "addbreaks" filter), then escape those breaks and double-escape the escaped markup-significant characters (with a bonus removal of invalid-in-XML characters, and an unneeded escape of apostrophes and quotes) with the "xmlescape" filter. That results in output which fits with Atom's type="html" content model, preserves possibly-significant whitespace, and permits clients to wrap as they see fit.

The perfect (read: enemy of the good) approach would be to write an xhtmlnl2brandp filter, which would convert double-newline separated blocks of text into paragraphs, single newlines into xhtml <br/>s, and probably add some ponies. Having seen such functions, in places like WordPress, I'll leave that for someone else to write.