reviewboard: send old review ids to server
When pushing changesets for review, the client may now optionally send
the review identifier previously associated with that changeset.

Obsolescence data is consulted to find reviews associated with precursor
changesets.

We also refactored the displayed data to be slightly more user friendly
and to cut down on hackery in testing land.