reviewboard: integrate Review Board server into tests
We now launch a local Review Board service instance when running the
Review Board tests.

Along the way, the logic for mapping commits to reviews was rewritten
to make it behave slightly better.

The tests now need rework to download results from Review Board to
ensure that the interaction actually does what we want it to do.

There is a test failure in test-obsolescence.t, presumably due to
a bug in commit<->review reconciling.

The daemon process running the Review Board server also appears to
get orphaned after the tests finish.

These issues will be addressed later.