Workaround to make sure navbar goes right above review details box.
There's no TemplateHook injection point that allows us to put content
between the review request banners (like the draft banner, for example),
and the review details box. We workaround this by using Javascript to
position the node on load. We can revert this once the appropriate
hook has been added to Review Board.