Commit all child review requests if squashed request is published.
As the summary says, this causes us to try to publish all child
review requests of a squashed request when the squash request is
reviewed. This is probably going to break if you don't have a
default reviewer configured, and you just go ahead and publish,
since Review Board expects those child requests to have reviewers
assigned.

This shouldn't change the behaviour of non-squashed review requests.