tests: support mocking Bugzilla server
We now have an extension that lies about local repositories being
Mozilla repositories. This allows us to test extensions that influence
interaction with Mozilla repositories.

We put this functionality to work by writing a quick test for bzpost.

Using mocks for Bugzilla is unfortunate. But it's easier than starting a
BMO server. We'll do that eventually.