firefoxtree: support serving firefox tree tags via wireproto
Servers can now advertise and serve a mapping of Firefox tree tags to
clients. When pulling, clients will now download this mapping and apply
the tags locally.

When applying pulled tags, the client will print a report on which tags
have been updated. This is similar to what happens when bookmarks are
pulled.