reviewboard: Bug 1035502 - Copy some p2rb extra data from the draft on publish. r=mconley
Now when publishing keys listed in DRAFTED_EXTRA_DATA_KEYS will be
copied from the draft to the review request object. This is important
for allowing a push to be completely rolled back before publish without
showing any user facing changes.

We also no longer store the commit id in the commit-id field and instead
store it in 'extra_data.p2rb.commit_id', which takes advantage of the
drafted extra data part of the patch.

I have not updated the tests to reflect these changes, as we will need
to update them quite a bit when the rollback functionality lands soon
after. The tests do not actually exercise much of the code anymore as
well because they never publish drafts. We'll want to write new tests
which actually publish along with fixing up the current tests.