Bug 1034171: Introduce the commits tab and restrict operations to the squashed review request. r=smacleod.
No edits should be made directly onto review requests. The only changes
that can be made are either via pushing (this will set the description,
commit IDs, repository, etc), and via the special RBMozUI view that allows
for batch assignment of reviewers and batch publishing.

Also, we only apply RBMozUI to pushed review requests.
Also, we set the reviewers on the squashed review request to be the union of the child reviewers.