rbbz/rbmozui: Update tests to pass after changes from bug 1034168 and bug 1035502 land. r=smacleod.
Bug 1034168 and bug 1035502 changed how almost everything works, and how things are
represented within Review Board with regards to the push workflow. That means that
certain changes that we tested for would only be visible after publishing the
squashed review request. We've modified the tests to account for this.

We also chose to disable a test (test-diffs) that doesn't appear to test
anything at this time. This test doesn't actually confirm there are parent
diffs, nor properly check what the base commit ID is.