Bug 1054517 - Port commit-messages hook tests to .t tests; r=edmorley
During the port, it was realized that the testing methodology in
runtests.py was flawed. Since commits were being stacked, the old
testing code only verified that any commit was bad, not necessarily
that the top commit was bad. This revealed that the hook was examining
the full commit message for bug and backout syntax, as opposed to only
the first line. The hook has been updated to make the tests pass.