Bug 1047516 - Send obsolescence markers as part of review request; r=mconley
Previously, clients consulted local obsolescence markers and attempted
to map old review requests to new commits. This worked in the simple
case. However, a big problem was the client's state getting out of
sync with the server's. To solve this state problem on the client would
require similar review request mapping code to live in the client.
This problem would be very difficult to solve.

This patch defers the responsibility of commit mapping to the server
and avoids the local review request sync problem entirely.

Instead of sending a mapping of new commits to review requests to
the server, the client instead sends the set of precursor nodes for
each reviewed commit to the server. The server then consults this
information and performs review request mapping. The server is
making all decisions about review requests and the client's code
path can remain simpler.