Bug 1065050 - Write to pushlog through extension API; r=glandium
This patch refactors the pushlog hook to write to the pushlog using the
API provided by the pushlog extension.

The error message for a locked database has been changed slightly to
avoid a leaky abstraction. It shouldn't matter: the important thing is
the message asking people to push again.

A new test for detecting a missing extension has been added.