Bug 1065050 - Rewrite pushlog as an extension; r=glandium
The existing code from the pushlog hook for establishing the pushlog
database and writing to it has been ported to a Mercurial extension.

The extension itself doesn't do anything except add a .pushlog attribute
on the repo class and define the class behind that attribute. Upcoming
changes will make the pushlog do things.