Bug 1065050 - Move pushlog tests into extension; r=bkero
Now that all pushlog functionality has been moved to the pushlog
extension, the tests have no place next to the hook. So, they've been
moved next to the extension.

There does remain one hook test, and that's for the hook being installed
without the extension. That test now exists in a new file.

This test can be deleted when the legacy hook is deleted.