pushlog: handle transaction rollback (bug 966545); r=glandium
With this patch, the pushlog now correctly handles transaction rollback.
If the transaction is aborted, the current sqlite transaction for the
pushlog insertion is also aborted and no data is retained.

This means that hook order or ^C during pushes will be irrelevant and
won't result in the pushlog getting in an inconsistent state with the
rest of the repository.