pushlog: don't record changes in pushlog unless they came from a push; r=glandium, r=bkero
The pretxnchangegroup hook runs whenever changesets are introduced to a
repository. This can happen during pull or commit, for example. We
update our hook to ignore changesets that come from unknown sources.

This is arguably the wrong thing to do, as it means we don't have
history for certain changesets. The robust solution would be to split
the changelog into local and remote components. That is out of scope...
for now.