pushlog: add better tests for HTML output; r=glandium, bkero
Our current test coverage of pushlog HTML is not great - we simply do a
sanity test. With this patch, we start verifying the returned HTML
content is what we expect.

To facilitate testing only a subset of returned content, the templates
were modified to make identifying pushlog entries from the XML output
a bit easier.