mozhg: copy profile finding functions from bzexport into mozhg.auth; r=smacleod
This patch starts a series of patches that extract the logic in
bzauth.py into mozhg.auth, where it can be more easily reused and
tested.

No copied code was modified as part of the move. It will be cleaned up in
a subsequent patch.

While we'd prefer to move the bzauth.py code explicitly, bzexport has no
test coverage. Subsequent patches appeared to break subtle behavior in
bzexport. So, we're making a copy so we don't risk any regressions to
bzexport.