mozhg: copy get_cookies_from_profile into mozhg.auth; r=smacleod
Our next step porting bzexport's auth code to mozhg.auth is to copy the
sqlite cookie.db reading. This patch does this. Copied code was not
modified. Style, etc will be fixed in a subsequent patch.

Again, we'd prefer to move the code. But bzexport is a fragile flower
and we don't want to risk breaking it.