docker: add a script to control Docker containers
The plan is to incorporate Docker into the automated testing
environment, including as part of Mercurial .t tests. For that, we'll
need a way to provision test-specific Docker containers.

This patch introduces a docker-control.py script for managing Docker
containers. It's a very high-level API on top of Docker's API to perform
very well-defined tasks.

As part of this change, the docker-py package and its dependencies have
been added to the virtualenv.

The new script is not yet considered production ready.