bzpost: convert tests to docker (bug 1078659); r=glandium
Now that we can run a local BMO instance via Docker, we convert the
bzpost tests to use it.

This initial patch does the conversion by setting up a BMO instance,
creating empty bugs, and changing the bugs referenced in the existing
tests. This results in hashes changing, which results in more test churn
than would be desirable.

Since we no longer use mocks.py as part of the test, we remove it and
the hack in local-mozilla-repos.py to have all requests-driven HTTP
requests be mocked, which mocks.py was the only consumer of.