docker: force mysql to listen on IPv4 and disable name resolution
Inspection revealed that MySQL was sometimes listening on IPv6. We don't
want IPv6 to enter the equation. Disable it by setting the bind_address
to 0.0.0.0, which will only use IPv4 interfaces.

Name resolution is also disabled because MySQL will attempt a reverse
address lookup without it. We want it using IP addresses everywhere
because that's what the permissions are defined as.