docker: wait for HTTP response, not socket
On some systems, facilities such as systemd will accept socket
connections for services that haven't come online yet. Therefore,
checking for socket connectivity isn't sufficient. So, we look for a
valid HTTP response instead.

We were seeing this on the Ubuntu-based Jenkins VM. It took many hours
of debugging to track this down.