docker: update Bugzilla Git repository during container startup
We want tests to be continuously running against a modern BMO code base.
So, we add some logic for updating Git to entrypoint.py.

Since updating Git can muck with applied patches, we moved the code for
patch application out of Dockerfile and into entrypoint.py

Since we want to be able to start containers when offline, this option
is behind an environment variable.