reviewboard: handle auth errors gracefully; r=smacleod
As part of implementing better auth support, it was realized that bad auth
credentials would result in an uncaught rbtools exception whose stack
would get printed to the client. This was not very helpful. This patch
traps that exception and turns it into a nice error code.

As part of this patch, we added an extra field to the protocol response
to indicate the client should raise an error.

This patch does not test cookie failures. That will be added in a
subsequent patch.