firefoxtree: support Mercurial 3.2 version of pull() (bug 1082937); r=glandium
Mercurial 3.2 (currently in @) removed localrepository.pull in favor of
using exchange.pull directly. We refactor our code to favor
exchange.pull as well.

As part of this refactoring, we had to update references to the repo
from "self" to an argument. We also had to add the block checking for
"is firefox tree" into pull() since we now wrap a global function, not
a per-repo one.