hghooks: delete unused prevent_unlabelled_australis_changes hook (bug 1075318); rs=jaws, edmorley
The prevent_unlabelled_australis_changes hook was added in bug 943486 to
help with the Australis transition. According to comment 42 and after in
that bug and bkero's forensics in bug 1075318, this hook is no longer
used and is thus being deleted.

This patch didn't technically receive a review. But it's deleting a
single, untested file. What could go wrong?