reviewboard: tests for automatic creation of users in Review Board; r=smacleod
When Review Board sees valid Bugzilla credentials for a user that
doesn't exist in the Review Board database, Review Board should
automatically add that user to Review Board. This test verifies that
happens during `hg push`.

There are 2 test variations, 1 for a user with an IRC nickname and 1
for a user without an IRC nickname.