reviewboard: add a test for existing users using a conflicting username; r=smacleod
If there are 2 existing Bugzilla and Review Board users and a user
changes their IRC nickname, Review Board should reassign the username to
the email+id form, just like it does if a new Review Board user is
created.

The code in rbbz for this case is a separate code path from the previous
test.