mqobs: delete experimental extension
The mqobs was always experimental. It wasn't feature complete and didn't
really work too well. The features are better implemented upstream
anyway. Lately, the extension starting breaking in Mercurial @ (soon to
be Mercurial 3.2). Since mqobs doesn't work and it has an unclear
future, we just delete it. We can always resurrect it if it is needed
later.