firefoxtree: prevent unknown reference to _updateremoterefs (bug 1085066)
A user reported that the extension was attempting to access an unknown
attribute on a repo instance. The backtrace indicated filtered
repositories may be involved.

The extra localrepository class for Firefox repositories provided
only a single method, which didn't need to be on a class. To match the
recent move of pull() into a module-level function, we move the
remaining class method into a module-level function and remove the
custom class. There should now be no issue accessing this function on an
unknown type.

Unfortunately, no test for the underlying issue is included. However,
the solution does treat the symptoms. There may be a bug still
lingering, likely around calling the firefoxtree code when it shouldn't
be.