reviewboard: use mach in rbmanage.py
This patch starts the process of converting rbmanage.py into a mach
script.

We set up a mach dispatcher in the script and port the "create" action
to a mach command.

Old actions are still processed in rbmanage.py. We will convert the
remaining actions into mach commands in subsequent commits.