reviewboard: add an add-reviewer mach command; r=smacleod
In order to test review integration with Bugzilla, we need a way to
specify a reviewer on a review request. We add a mach command to
facilitate this.

The hack to query the user API to cause the user to be created in Review
Board is extremely hacky and is a horrible user experience. We should
definitely look into intercepting the strings when the review request is
updated and trigger the necessary Bugzilla lookup so Review Board
doesn't complain about unknown users when you submit a username that
isn't defined in Review Board yet.