reviewboard: properly set review flag requestee (bug 1085698); r=smacleod
This patch fixes an apparent regression in the refactoring of Review
Board username handling.

The code for adding a review request flag on a Bugzilla attachment was
using the Review Board username for the Bugzilla user. This is obviously
wrong. We update the code to use the email address, which is the
identifier the Bugzilla API is looking for.

A test verifying that review flags are added during review request
publishing has been added. The test fails without the changes in
rbbz.extension. Note that a new review flag appears. This demonstrates
that the r? flag wasn't being set before because of this bug.