mozreview: add high-level APIs and tools for controlling MozReview instances
To make it easier for individuals to launch MozReview instances, we add
a high-level MozReview class for controlling isolated and local
MozReview instances. There is a single start() API where it will
spin up Dockerized BMO, a local Review Board server, and a local
Mercurial server.

A "mach" driver has been added to the root of the repository. It
contains commands for manipulating MozReview instances.