reviewboard: use email address for requestee when updating review flags (bug 1087617); r=mcote
Like bug 1085698, the conversion of Review Board usernames from emails
to usernames resulted in Review Board sending a username (not email) as
the requestee parameter when updating attachment flags. This was causing
Review Board to never find the existing flag, which resulted in the
creation of a duplicate flag.

The fix is simple: use the email address for requestee.

A number of tests were updated. In hindsight, those tests demonstrated
the bug, but nobody noticed. Oops!