jenkins: do not destroy VM if last build failed
e60b276cd850 introduced code to have the Jenkins job destroy the VM if
the last build failed. This was a temporary and very hacky workaround to
deal with inconsistent and non-robust VM state. Things are in a much
better state now. So, we revert this commit and no longer destroy the VM
after a failed build.