hghooks: remove prevent_broken_csets hook (bug 1075275); r=glandium
The prevent_broken_csets hook is currently only deployed on a single
repository (releases/mozilla-aurora). That itself is probably a bug (it
should have been more widely deployed).

Since prevent_broken_csets's lack of test coverage is blocking the
upgrade of hgssh to 3.1 *and* since the newer (and still not deployed)
changelog_correctness hook (from bug 962300) is much better, we've
decided to kill this under-used, less powerful hook. The plan is to
delete this hook, unblocking the hgssh upgrade, and to eventually roll
out a proper form of changelog_correctness to a wide base of
repositories.