reviewboard: verify bugs before reviews are created (bug 1034178); r=smacleod
Review Board doesn't currently handle failures during publishing too
well (bug 1088823). As a result, the error due to unknown or
confidential bugs isn't being displayed via the web interface. This
results in a horrible user experience.

Between this and the principle of "fail fast," we introduce
functionality to the pushreviews Mercurial server-side command handling
to verify that the bug referenced by a review ID is accessible and is
appropriate for receiving review requests.

We rename test-publish-errors.t to test-push-invalid-bug.t to reflect
the new nature of where the check is enforced.