testing: add vcttesting to virtualenv
The support modules in the vcttesting package are becoming increasingly
popular. This directory currently must be manually added to sys.path.
This is annoying.

This patch adds packaging to vcttesting so it is available in our
virtualenv.

Scripts manually adding the testing directory to sys.path have been
updated to no longer do this.