testing: add Docker build environment for eggs (bug 1090575); r=mconley
The Review Board eggs are annoying to build. We need Python 2.6 to match
production. We need some node.js packages to produce the static assets.

This patch adds a Docker environment that provides a build environment
for the Review Board eggs. It goes one step further and provides a
high-level command to start the container, generate the eggs, retrieve
them, and save them to local disk.