reviewboard: obsolete review attachments when discarding review request (bug 1004835); r=smacleod
Discarding a review request effectively says "we don't care about this
review request any more." Obsoleting an attachment in Bugzilla
effectively means the same thing.

This patch adds functionality to rbbz that will obsolete Bugzilla
attachments for review requests that are obsoleted.

The implementation is not terrific yet: reopening a discarded review
request will create a new attachment instead of removing the obsolete
flag from the identical existing attachment. This will be addressed in a
subsequent patch.