reviewboard: don't always import json (bug 1092476); r=mconley
The json module is not available in the Windows Mercurial distribution.
Having |import json| at the module level causes the reviewboard
extension to fail to load, preventing use completely.

In this patch, we move the json import into the the function it is
needed in. We also add basic test coverage to ensure that "banned"
global imports aren't present after the extension is loaded.