Bug 1097736 - rbmozui: Can't assign reviewers to a review request. r=smacleod.
Comparing users for sorting in the autocomplete popup was not taking private
user accounts into, er, account. Because fullnames are not exposed for private
accounts, this caused comparing to sometimes fail with certain results. We now
handle the case where fullnames do not exist by setting non-existant fullnames
to the empty string.