hghooks: make push_printurls.py output hgweb URLs for Try pushes too (bug 1096917)
Currently hgweb URLs are only output for non-try pushes, with Try pushes instead
linking to Treeherder/TBPL. However it's still useful to output the hgweb URL
for try, since devs wanting to check their commits can do so without having to
go via Treeherder (and thus having to wait for it to ingest the push).

This patch means that we output hgweb URLs for all pushes. The order of the hook
is changed, since it makes more semantic sense for the revision link to be
output before linking to results based on that revision.