Bug 1104390 - Remove legacy web-based repository visualizer; r=bkero
There was old code in the pushlog extension for doing visualization of
repositories using AJAX and other nifty web visualizations. This
interface was all built on top of a set of custom hgweb commands that
emitted JSON.

AFAICT, this code is dead and broken. It is thus being removed.

This commit is logically a backout of dddf4b816e51, plus all the changes
since made to the files it has introduced.