pushlog: add last push ID to pushlog response (bug 1065771); r=edmorley
Clients want a way of polling for changes since they last fetched. We
want them to do this by looking at the most recent push id and fetching
all changes starting from that ID + 1.

Before this patch, clients fetching this way had no way of knowing if
the pushlog database was reset because fetches with last push id + 1
would return an empty set until the new database caught up to the old
one. This could take months!

By adding the last push ID to the response, clients can store and
blindly use this value as an offset into the next query (assuming
this value is <= the highest seen value + 1).