pushlog: support pulling pushlog data over wire protocol; r=glandium
This patch introduces a "pushlog" wire protocol command that returns
information on pushes.

The pull operation is overloaded in the client to fetch this pushlog
data, if available. The fetched data will be added to the local pushlog
database as if a push occurred to the local machine.

Currently, the pushlog pulling assumes a unidirectional flow with a
single source of truth. It thus only really works for master-slave
replication. Pushlog data is generally useful and giving clients
access to pushlog data (possibly via the same extension) is a noble
goal. However, it is not an immediate goal.

We should consider having a separate mode for "advanced" pushlog use
cases (including multi remote) or we should consider splitting the
pushlog extension into different extensions, each optimized for a
particular use case.