firefoxtree: allow aliases to work on push (bug 1105833); r=bkero
With this patch ``hg push <tree>`` now works.

The UI is a bit wonky. Because we're doing path expansion before the
command is invoked, the output from ``hg push`` prints the expanded
URL rather than the alias the user entered. This is contrary to how
Mercurial normally works. But, that's a small price to pay for
convenience.

Now that push aliases are working, the test mocking code has been
changed to always use ssh for pushes and to disallow pushes via http,
just like production.