reviewboard: perform bug comment & flag update as a single action (bug 1098093)
Prior to bug 508541 being fixed, Bugmail for bug comments was not correctly sent
if the comments were made at the same time as the flag change. As such, a
temporary workaround was implemented, that performed the comment as a separate
action.

Now that bug 508541 is fixed, we can go back to performing these changes as a
single action, to reduce Bugspam & noise in inline bug history, as well as mean
that people who filter certain classes of Bugmail don't get a confusing view of
changes that occurred.