firefoxtree: allow pushing multiple heads to non-canonical repos (bug 1115199); r=glandium
firefoxtree changes the behavior of `hg push` to only push '.' if a
Firefox repository is involved. Before, this check was based on "is the
current repository a Firefox repository." This was contraining because
some people would want to push their Firefox repository to e.g. a
bitbucket or user repository.

After this patch, the behavior is only to push '.' by default if the
remote is a canonical Firefox repository. This accomplishes the goal of
the feature (preventing accidental multi-head pushes to canonical repos)
without limiting valid workflows.