pushlog: ensure usernames are UTF-8; r=glandium
We want pushlog to fail fast if usernames aren't valid UTF-8.

I attempted to add a test for this. However, I could not coax
shell/Python into exposing binary strings to os.environ. I was able to
set a bash variable with binary content. However, when it got to
Python's os.environ, the backslashes for \xHH were escaped as \\xHH.