pushlog: do not attempt to create pushlog for read-only operation (bug 1116328); r=sfink
If a pushlog operation is read-only and a pushlog database does not
exist, there is no need to attempt to create a pushlog database.
Attempting to do so may encounter permissions errors (as seen in the
wild).

This patch changes behavior such that read-only operations will not
attempt to create the pushlog database.