hgtemplates: move "age" class in shortlogentry; r=bkero
The infinite scrolling code in hgweb performs date munging on shortlog
elements with the "age" class. Having both the node and date in the same
node tree confuses the munging and results in poorly concatenated text
content being assigned to the node. Essentially, all our pretty
formatting is lost.

We work around this by assigning the "age" to a more inner element.