pushlog: wrap exchange._pullobsolete (bug 1118619)
Mercurial 3.3 refactored the transactions API and removed our previous
hook point: exchange.pushop.closetransaction. The pull code in both
versions is essentially:

  _pullbookmarks()
  _pullobsolete()
  closetransaction()

This patch changes the hook point from inside and before
closetransaction to inside and after _pullobsolete.

The extension still fails to write new data properly in 3.3 due to
failures in transaction monkeypatching. This will be fixed in a
subsequent patch.