rbbz: call appropriate API for parsing xmlrpc response (bug 1118650); r=smacleod
We were manually performing low-level xmlrpc parsing work. This patch
switches to use an appropriate API on xmlrpclib.Transport.

I noticed a problem when I updated Bugzilla to a newer commit. HTTP
responses contained Accept-Encoding: gzip and we were attempting to send
a buffer with compressed data to an XML parser. I thought it was strange
that standard HTTP encoding headers weren't automatically being obeyed,
so I dug into the source code and found an API that automatically
handles the Accept-Encoding header.