tests: dump more bug state; r=sfink
Upcoming tests will need to verify that currently untracked bug
properties are set as expected. We add these fields to the
|bugzilla dump-bug| output.

The extra output is slightly annoying in cases. But, having more output
can be a useful signal and cause a human to go "that doesn't look
right." So, I think having all the output all the time is a good thing.