bzexport: support for using a requests.Session for making requests; r=sfink
Upcoming patches will introduce usage of the Bugzilla REST API into
bzexport. In order to support this, we need to handle token
authentication. We pave the road for this by introducing a mechanism to
obtain a ``requests.Session`` instance to be used to talk to Bugzilla.
The session will automatically issue the appropriate cookies and/or
parameters to facilitate authenticated communication.