bzexport: use mozhg.auth.getbugzillaauth; r=sfink
We have a high-level API in mozhg.auth for obtaining Bugzilla
credentials in a Mercurial manner. Switch to it.

Note that this change is technically backwards incompatible because
behavior is slightly different. For example, mozhg.auth will search
*all* profiles, not just the default profile. If people want the old
behavior, we can always add a flag to mozhg.auth to provide it.

This patch undoes the bulk of our earlier patches. But hopefully the
small, self-contained patches made it easier to review.

While we were here, we removed some unused imports.