rbmozui: properly bump the p2rb.reviewer_epoch on the root review request. r=smacleod.
We bump the epoch (a counter) on the root review request when we change the
reviewers on child review requests. We do this so that we can make the root
review request have a valid change to publish - otherwise, Review Board
complains that nothing about the root review request has changed.

We were doing this before, but doing it incorrectly - we need to pass a
jsonFieldName to the draft editor. This seems redundant, but it's reality.