mozreview: Only send published messages for parent review requests (Bug 1124259). r=gps
When support lands in RB core for multi-commit review requests we will
only ever have a single entity being published, so we choose now to send
a message for just the parent review request. If a consumer would like
to perform actions on the specific commits they will be able to extract
that information from the parents data.