mozreview: Introduce an API to create reviews in one request (Bug 1124502). r=gps r=mcote
Add a new API endpoint for automation to post reviews using a single
HTTP request. This is required for concurrent reviews since a single
user can only craft one review at a time using the traditional APIs.

Using this endpoint is also much faster as it requires only a single
roundtrip in contrast to 2+n roundtrips where n is the number of
comments.