testing: remove containers faster
When stopping MozReview, we first issued a graceful stop then a remove
container. Since we're removing the container, we don't preserve data.
So, the time it takes to allow the containers to shut down gracefully is
wasted. We switch to just force remove the container, which kills the
container.

On my machine, stopping containers now takes ~2s instead of ~10s!