testing: issue warning when Docker cert error is encountered
This is a very weird issue that seems to reproduce in Python 2.7.9 +
docker-py + boot2docker 1.4.1. mcote and I both can reproduce it. The
underlying failure is within the ssl C extension. It smells like Docker
isn't producing certificates properly or a regression in 2.7.9 or the
inability of requests/urllib3 to set up the verification environment
properly. Whatever the root cause, this is extremely annoying. We should
alert the user.