testing: separate global config in MozReview Mercurial server
We refactor the configuration of the Mercurial server in MozReview
instances such that the global configuration is different from the web
configuration. This simulates how things work in production, which a
global hgrc (typically /etc/mercurial/hgrc) provides global defaults and
a web config file provides the web configuration.

These don't strictly need to be separate files. But it helps keep logic
sorted out. And considering Mercurial's behavior of only reading a
whitelist of options from web.conf files, it makes sense to keep a wall
between them to avoid ambiguity.