testing: serve multiple repositories from commonenv test server
We are preparing to use mozreview in the test environment. It creates a
Mercurial server that supports multiple repositories as opposed to a
single repo server.

This patch converts the MozReview test environment to use a server setup
that supports multiple repositories. This will make the switchover to
mozreview less painful.