testing: properly daemonize Review Board server process
Our previous implementation of the double fork was incorrect. We were
performing a sys.exit() in the parent process and resuming control from
the child process. When executed in the context of a shell, this caused
the shell to proceed with subsequent commands immediately. This was
causing race conditions between MozReview.start() writing state and
other processes accessing that state.