testing: use concurrent.futures to clean Docker state after test run
We were previously iterating over all containers and issuing a stop
request. This was inefficient because it was sequential and we were
waiting on containers to gracefully shut down, only to remove them. We
improve the situation by using concurrent.futures and forcefully killing
containers.

If we have a lot of orphaned containers after test execution, we now
clean up significantly faster.