hgtemplates: remove websub where buglink is used (bug 1123992)
The "websub" filter is currently incompatible with the "buglink" filter
because the global hgrc as an "[interhg]" (equivalent to "[websub]")
entry for rewriting bug numbers into URLs. This was resulting in some
double encoding and malformed HTML.

A proper fix is probably to kill the buglink filter and switch to
websub. But that requires Puppet changes and a more invasive coding
change.