hgrb: Create functions for code that is shared between pushing reviews and reserving review ids (bug 1128977); r=gps
There is some shared code between pushing a review and reserving review ids for
a push. This breaks this code out into functions both for ease of reuse and to
improve readability. In the future we should work to minimize the duplicated
work between reserving a review and pushing it, but for now, we can't assume
that a client has first reserved review ids prior to attempting to push.