pash: reconcile changes with upstream (bug 1089385); r=bkero
The pash version imported into version-control-tools was not the one
currently deployed on the SSH hosts. One of the differences is the repo
root path. So, we change /repo_local to /repo.

There was some additional changes upstream to actions when user repos
are created. We add those.

The "repo-group" command was added. We also add it.