mozext: fetch pushlog automatically during pull
This is a low-hanging fruit that's now possible without additional
server round trip since the pushlog extension is advertising a
capability on the server. This should help make the problems with
the pushlog being out of sync (because you need to manually run
`hg pushlogsync`) a thing of the past.