hghooks: remove duplicated logic in comm-central treeclosure hook (bug 1138089)
Now the common treeclosure code has been refactored out, we can make
treeclosure_comm_central use it to avoid duplication. To simplify the
implementation, we don't preserve the exact message that was shown previously,
but instead use the treestatus name - which for comm-central is the repo name
with the app name appended. To me this actually makes more sense, since it
directly matches the name that will be shown on treestatus.mozilla.org, so
easier for people to cross-reference.