testing: adjust tests to be compatible with Mercurial 3.3
Mercurial 3.3 changed output of the graft, rebase, and strip commands
slightly. There also appears to be a difference in reporting on file
counts during pushes in Mercurial 3.3.

This patch quiets some commands and adds globbing elsewhere so that
tests pass on Mercurial 3.3.

According to Jenkins, we still have some tests with different output on
Mercurial 3.0 and 3.1. But this patch addresses most of the
inconsistencies.